Timestamps with logging and mongo server change
I added log4js to the package. This is to hopefully solve the issue that
the server has sometimes been shutting down (actually, this is the
reasoning behind all the improved logging). Mainly, log4js is being used
because it timestamps everything (including console.log). Additionally,
it's pretty useful if we start to use it instead of console.log.
The mongo server has been moved from the temporary server to
gamez.animatron at the standard port.